requirements.txt
# ===========================================================================
# Vantag Platform — Python Requirements
# ===========================================================================
# Install with:
#   pip install -r requirements.txt
#
# On Jetson (ARM64), TensorRT and PyCUDA are provided by the L4T base image
# and must NOT be installed via pip.  The TRT/CUDA stubs below are
# conditionally imported at runtime; they are not listed here.
# ===========================================================================

# ---------------------------------------------------------------------------
# Web framework & API
# ---------------------------------------------------------------------------
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-multipart>=0.0.9     # multipart form uploads (watchlist images, etc.)
aiofiles>=23.2.1             # async file I/O for snapshot serving

# ---------------------------------------------------------------------------
# Computer vision & ML inference
# ---------------------------------------------------------------------------
opencv-python-headless>=4.9.0   # headless (no GUI); compatible with Jetson
numpy>=1.26.0
ultralytics>=8.2.0              # YOLOv8 detection + pose
supervision>=0.20.0             # annotation utilities (ByteTrack, zones)
onnxruntime>=1.18.0             # CPU/GPU ONNX inference (face recognition)
insightface>=0.7.3              # face detection + ArcFace embeddings

# ---------------------------------------------------------------------------
# Spatial & zone analysis
# ---------------------------------------------------------------------------
shapely>=2.0.3                  # polygon zone membership (dwell, loitering)

# ---------------------------------------------------------------------------
# Predictive scoring
# ---------------------------------------------------------------------------
lightgbm>=4.3.0                 # gradient-boosted predictive risk model

# ---------------------------------------------------------------------------
# MQTT
# ---------------------------------------------------------------------------
paho-mqtt>=2.0.0

# ---------------------------------------------------------------------------
# Reporting & document generation
# ---------------------------------------------------------------------------
reportlab>=4.2.0
pillow>=10.3.0                  # image compositing for PDF snapshots

# ---------------------------------------------------------------------------
# Security & cryptography
# ---------------------------------------------------------------------------
cryptography>=42.0.0            # Fernet encryption for face embeddings
PyJWT>=2.8.0                    # JWT tokens for API authentication

# ---------------------------------------------------------------------------
# Configuration & serialisation
# ---------------------------------------------------------------------------
PyYAML>=6.0.1                   # cameras.yaml / config loading

# ---------------------------------------------------------------------------
# Logging & observability
# ---------------------------------------------------------------------------
structlog>=24.1.0               # structured JSON logging

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest>=8.2.0
pytest-asyncio>=0.23.0          # async test support for FastAPI
httpx>=0.27.0                   # async HTTP client for TestClient

# ---------------------------------------------------------------------------
# SaaS multi-tenant additions
# ---------------------------------------------------------------------------
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
redis>=5.0.0
razorpay>=1.4.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sendgrid>=6.11.0
email-validator>=2.1.0
python-dotenv>=1.0.0
